downgrade to 8.5.0

+++ src/train.py
@@ -11,29 +11,33 @@ role = "arn:aws:iam::278088188282:role/service-role/AmazonSageMaker-ExecutionRol
 region = sess.boto_region_name
 bucket_name = "popiol.daytrader-master-quotes"
 bucket = s3.Bucket(bucket_name)
-files = [x.key for x in bucket.objects.filter(Prefix="csv_clean/date=202105")]
-print(files[:5])
-quotes = None
 
-for file in files[:400]:
-    quotes1 = pd.read_csv(bucket.Object(file).get()["Body"])
-    quotes = quotes1 if quotes is None else pd.concat([quotes, quotes1])
+if False:
+    files = [x.key for x in bucket.objects.filter(Prefix="csv_clean/date=202105")]
+    print(files[:5])
+    quotes = None
 
-quotes2 = quotes.rename(columns={"quote_dt": "start"}).drop(
-    columns=["low_price", "high_price", "row_id"]
-)
-grouped = quotes2.groupby("comp_code")
-start = grouped["start"].min()
-target = grouped["price"].agg(lambda x: x.tolist())
-train = pd.DataFrame({"start": start, "target": target})
-print(train.head())
+    for file in files[:400]:
+        quotes1 = pd.read_csv(bucket.Object(file).get()["Body"])
+        quotes = quotes1 if quotes is None else pd.concat([quotes, quotes1])
+
+    quotes2 = quotes.rename(columns={"quote_dt": "start"}).drop(
+        columns=["low_price", "high_price", "row_id"]
+    )
+    grouped = quotes2.groupby("comp_code")
+    start = grouped["start"].min()
+    target = grouped["price"].agg(lambda x: x.tolist())
+    train = pd.DataFrame({"start": start, "target": target})
+    print(train.head())
 
 train_file_key = "train.jsonl"
 train_file = f"s3://{bucket_name2}/{train_file_key}"
-with open("tmp.jsonl", "w") as f:
-    train.to_json(f, orient="records", lines=True)
-bucket2.upload_file("tmp.jsonl", Key=train_file_key)
-os.remove("tmp.jsonl")
+
+if False:
+    with open("tmp.jsonl", "w") as f:
+        train.to_json(f, orient="records", lines=True)
+    bucket2.upload_file("tmp.jsonl", Key=train_file_key)
+    os.remove("tmp.jsonl")
 # train_file_key = "train.parquet"
 # train_file = f"s3://{bucket_name2}/{train_file_key}"
 # with open("tmp.parquet", "bw") as f:
@@ -63,6 +67,6 @@ estimator.set_hyperparameters(**hyperparameters)
 
 print("Fitting...")
 
-estimator.fit(train_file)
+estimator.fit(inputs={"train": train_file})
 
 print("Model created")
